Re-enable tor install for monitor server
Was going back and forth on this decision, I've changed my mind. We
should always install tor on both servers mon and app. Thats the current
status quo and I'm not ready to make that unanimous decision without
going through the team and getting a better implication of that change.