Add detection when reboot needed for iptables
Previously system would not reboot as needed when iptables changes were
made after the initial install. With the updated logic when iptables
changes, we drop a file on disk indicating a reboot is needed. The
reboot task which is wayy further down the line sees that file and
reboots as necessary.